Move some update params here from fedora_openqa schedule.py
We had code in fedora_openqa schedule.py to set some params for
update tests depending on the flavor. I don't remember why I
ever wrote it that way, but that's an odd way to do it, the
obvious way to do it is just to have them here, like we already
have several others and like we do it for composes. This will
help with https://progress.opensuse.org/issues/119200 because
the params set by fedora_openqa schedule.py go in the request
URI but ones set here don't.

Signed-off-by: Adam Williamson <awilliam@redhat.com>