Move some tap job groups to new 'tap2' worker class
This is intended to allow us to split the workload across
multiple tap workers, as right now, sometimes we wind up in a
situation where all non-tap jobs are done and the non-tap
workers are sitting idle, but the poor tap worker has a backlog.
This way the tap jobs for updates can be split across two
worker hosts, which should help out.

Signed-off-by: Adam Williamson <awilliam@redhat.com>