Add `--squashfs-only` to lorax args
We started using this in real composes a year or two back, so
openQA should do the same. It drops the nesting of an ext4 fs
image inside a squashfs image, just using a single squashfs
image instead. This results in smaller images - missing this
is why the images built by openQA were coming out larger than
the real ones.

Signed-off-by: Adam Williamson <awilliam@redhat.com>