Just remove abrt_runs_found_problem needle
it's really just a dupe of the -problems needles, it turns out,
Lukas was reinventing that wheel. He had to add another one
today because I broke the JSON in this one when I was simplifying
it yesterday, but I think this one on the new -problems needle
are really just dupes.

Signed-off-by: Adam Williamson <awilliam@redhat.com>