rpmostree_overlay: try to cancel running operations
I've seen a few cases of this test failing because there was
some running operation when it tries to do `rpm-ostree install`.
I think this is GNOME Software checking for updates or something,
I've seen it on my own Silverblue install too. Let's just throw
some `rpm-ostree cancel`s at it and hope that helps.

Signed-off-by: Adam Williamson <awilliam@redhat.com>