Unset 'last time notification shown' setting for KDE (#2178311)
It looks like the desktop_notifications postinstall test on KDE
on F38 is failing currently because the notification is being
shown during the install_default_upload test that precedes it,
so KDE decides not to show it again yet. So, unset the setting
that stores the timestamp of the last time the notification was
shown. This is similar to a thing we already do for GNOME above.

Signed-off-by: Adam Williamson <awilliam@redhat.com>