Tweak LVM partition scheme needles a bit
In https://openqa.fedoraproject.org/tests/1938847 , we wound up
doing an LVM thinp install when we meant to do a regular LVM
install, because LVM was already highlighted (for some reason)
in the scheme list, and the "LVM" needle is narrow enough that
it matched on the start of "LVM Thin Provisioning".

To avoid this, we make the match area in the existing needle
wider so it can't match on "LVM Thin Provisioning", and add an
alternate needle for LVM when it's highlighted.

Signed-off-by: Adam Williamson <awilliam@redhat.com>