Drop syslinux handling from do_bootloader
F36 is now EOL, and from F37 onwards, grub is the bootloader in
any situation where it actually matters to do_bootloader (which
is only when we're editing parameters). We do still use syslinux
in the PXE tests on x86_64 BIOS, but we don't edit the parameters
in that case.

Signed-off-by: Adam Williamson <awilliam@redhat.com>