Adjust modularity_tests to enable the modular repos at the start
As of yesterday's Rawhide, the modular repos are not installed by
default, so of course all the modular tests fail. So, install
the repos before running the tests.

This isn't conditionalized on release version as I don't think
we ever run this test on anything other than Rawhide and
Branched.

Signed-off-by: Adam Williamson <awilliam@redhat.com>