Update the Gnome-Text-Editot test.
The editor started to show spell-checking that would require a lot
of new needles to be created. Theredore, we set the language to
English to stop showing the spelling mistakes in aaa_setup.pm

Also, the application started to have problems with getting correct
focus, so we want to click into the text before the status gets
recorded.