Update several needles for GNOME 45
The 'global' app menu is no longer shown, which affects the
apps_run_terminal needle (it will affect many others for the
compose tests, but this is just for updates tests). It seems we
sometimes have the cursor over the Activities button and it
looks slightly different in that case, so add a variant needle
for that. Finally, the password input box in GDM is no longer
highlighted in blue as it probably should be to indicate that
it's active, so add a workaround needle to handle that.

Signed-off-by: Adam Williamson <awilliam@redhat.com>