GDM language login needles: Rename, move and mark as workarounds
These are all the new needles for the 'no blue outline on
password entry box any more' issue. The naming was a bit non-
standard and they were not in the right directories, and they
need to be actually tagged as workarounds, like the English
one is. Also, add the correct issue link to the workaround note
for the English one.

Signed-off-by: Adam Williamson <awilliam@redhat.com>