Update various needles for EOG -> Loupe transition
Loupe replaced EOG as the image viewer in F39 and Rawhide. We
need to update several needles for this: its icon is different,
the change affected the appearance of the Utilities submenu
icon, its main UI looks different, and it uses a light colored
file chooser which requires some variant needles for the Maps
export image test.

Signed-off-by: Adam Williamson <awilliam@redhat.com>