Effectively skip anaconda_help on F39+ Workstation live for now
There isn't any Help in the webUI installer, really. In future
it's meant to have more 'contextual' help - little question
mark icons next to elements of the UI that explain them when
clicked on - which we could implement tests for, but this isn't
done yet.

Let's not skip scheduling the test entirely, because we can
still run it on F38 respins, and we may be able to implement
testing of the contextual help in future. So let's just soft
fail and return immediately. If we get to F39 stable without
the contextual help being implemented (or it turns out not to
be testable within the confines of this design), we can skip
scheduling the test on webUI images entirely.

Signed-off-by: Adam Williamson <awilliam@redhat.com>