Adjust konversation test to run neochat on Rawhide
KDE replaced Konversation (IRC client) with Neochat (Matrix
client) in Rawhide. As the replacement isn't done in F39 we can't
just switch the test out, we have to handle both, so for now,
let's have the "konversation" test run neochat on Rawhide.

We can't really proceed through neochat's first run wizard as
it needs a Matrix account name and password and we don't want
the hassle of handling a secret just for this, so we'll just
quit out once we see it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>