gnome-panel: set update notification timestamp early
If we do it after forcing the clock to 18:30, doing it takes so
long we wind up at 18:32, and that kinda sets the rest of the
test out of whack (we have several needles that assume we start
at 18:30 or 18:31 after the snapshot). So let's set the update
timestamp *before* we force the clock. This will mean it's waay
in the future after we force the clock, but that should still
do the job of avoiding notifications showing up, I hope.

Signed-off-by: Adam Williamson <awilliam@redhat.com>