Tweak update download needle to try and avoid upgrade match
In Fedora 38, now F39 is available, there's a Download button
for upgrade to F39 as well as a Download button for F38 updates.
Unfortunately they're nearly identical, and openQA is matching
on the upgrade one, not the update one. This stretches the match
area off the top of the button to include its different
background, which should make the match unique and I hope also
work for F37 (where the button is a different width, so we can't
easily go off the side). Will tweak if this doesn't work.

Signed-off-by: Adam Williamson <awilliam@redhat.com>